Add desktop fullscreen option in menu and decouple aspect ratio from window resolution.
Desktop Fullscreen:
It was previously possible to run ironwail in desktop fullscreen mode, but only by manually setting a cvar. This change adds the ability to pick this display mode within the video settings menu.

Decoupling Aspect Ratio:
Previously in ironwail the window/fullscreen size always dictates the aspect ratio tha the game is renderred in. This is fine for windowed and exclusive fullscreen where you can set the specific resolution you want, however with desktop fullscreen, the window must be the same size that your operating system is outputting. Decoupling the aspect ratio allows the user specify a specific maximum aspect ratio to use in this circumstance. For example 4:3 on a standard 16:9 display, or 16:9 on an ultrawide display. The new vid_maxaspect cvar default to 0, which means auto (match window/fullscreen resolution).